+++ integrations/opendataloader_pdf/src/haystack_integrations/components/converters/opendataloader_pdf/__init__.py
@@ -1,0 +1,11 @@
+# SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from haystack_integrations.components.converters.opendataloader_pdf.converter import (
+    OpenDataLoaderConverter,
+)
+
+__all__ = [
+    "OpenDataLoaderConverter",
+]